Add Support for Floating Point Textures on Mobile Platforms. (#6264)
* Add Support for Floating Point Textures on Mobile Platforms.

We currently restrict Android and iOS to NEVER support
`SurfaceFormat` values of `SurfaceFormat.Single` etc.
Even if the device supports it either directly or via
and OpenGL extension.

This commit reworks the GraphicsCapabilities and GraphicsExtensions
to use the queried extensions to figure out if we support the
`SurfaceFormat` requested. If it doesn't then we throw a
`NotSupportedException` as we current do. If we do however then
we will run as normal.

Also fixed up a few hardcoded `PixelFormat` values so they have
proper enumerations.

* Fixed Desktop Detection

* Added a decent error message

* add more extensions

* add more extensions

* Updated to check for Normalized formats

* GLES 3 supports them all

* Fixed typo

* Updated GL checks

* Allow Instance Rendering on GLES

* Allow Instancing on mobile GL plaforms if it supports it

* fix error

* Fixed up comments

* Updated based on feedback